src/ints/bios.cpp: Valgrind shows "use of uninitialized variable" which at first looks spurious (it's pointing at a constant+(variable*2) expression passed to mem_readw), but at closer examination, comes as a result of the function's declaration of int port; being overshadowed inside the for() loop by Bitu port, in a manner that assigns the value of mem_readw(port) to port, like this:
Bitu port = mem_readw(constant+(port*2));

The guy who wrote this code, I think, thought that the compiler would
declare "port" for the inner scope, then use "port" from the outer scope
and multiply by 2 and add to the const expr and assign to the "port"
variable on the inner scope. Instead, the "undefined" error happens
because GCC interprets it as taking the (undefined) value of "port",
multiply by 2, add to constexpr, and assign back. Thus, the undefined
value warning.