updated notes and bugs regarding demoscene entry "Majic 12: Show" and why and when it triggers a "MCB list is corrupt" error message in DOSBox and how to reconfigure your dosbox.conf to work around it.
It turns out the demo (perhaps the graphical effect routines) clears and
overwrites memory in the 0xC0000-0xCFFFF range. If DOSBox has any
private areas or upper memory blocks there, the data there will be lost.
If upper memory blocks exist up there, they will be lost and DOSBox will
E_Exit() with a message stating that the MCB list is corrupt. When I
added code to dump the MCBs, I saw that when this happens, the MCBs
below the 640KB boundary are OK, the MCBs in the upper memory are zeroed
out.