bugfix: VS 2015 C runtime functions open/lseek/stat now enforce refusal to handle 2GB or larger files (in the same way Linux has already been doing for a long time). It turns out the inability to IMGMOUNT large files is tied to the use of stat() at one point. Using a #define to mask stat vs stat64 vs _stati64 and the stat struct resolves the issue.
bugfix: IMGMOUNT sizes[] array is pre-initialized to zero
now instead of letting random stack garbage determine disk
geometry in cases where -sizes is not given and DOSBox is
unable to autodetect geometry. Added code to fill in the
missing geometry with a reasonable guess. The guess is
reasonable enough to get Windows 98 SE to boot from an 8GB
disk image, at least.