ems emulation bugfix: EMS was allocating memory pages with the "sequential" parameter == false. this means there is the possibility that when VCPI emulation set up the protected mode structures there was possible memory corruption due to the assumption the pages were sequential.
ems hack: a new option has been added to specify that the "system
handle" should be allocated on an even megabyte if possible. if set,
EMS/VCPI system structures should be fully usable regardless of the A20
gate. however the hack also causes some extended memory fragmentation
and reduces the chances a DOS program will be able to allocate a large
contiguous region of extended memory, so it is not enabled by default.