GUS programming mistakes fixed in emulation.
Bit 0 of the reset register is clared to reset, set to enable normal
operation. Prior to the fix, DOSBox interpreted setting the bit as
reset, which is wrong.

Bits 0-7 of the 16-bit gRegData variable were interpreted as the
contents of the 8-bit reset register, when it should have been looking
at bits 8-15. On the GUS, I/O port 3X4 is a 16-bit I/O port for the full
16-bit register while 8-bit wide registers are supposed to be
accessible from 3X5 as the upper 8 bits of the full 16-bit gRegData
value.

GUS initialization should be setting the reset register to 0x7 (writing
0x700 to gRegData) which is interpreted as (non-reset, DAC enable, master
IRQ enable).

GUS IRQ checking should be asserting and de-asserting the IRQ, not just
asserting the IRQ and leaving it on. It should also check the master IRQ
enable bit given in the reset register.

Additional reminders to myself were added for further accuracy fixes of
Gravis Ultrasound emulation.

So far, nothing seems to have a problem with these emulation fixes. Only
thing left to test is the official Gravis Ultrasound DOS utilities.