3dfx/voodoo OpenGL bugfix.
[from ghoul8765]
At Vogons, there is a workaround patch to avoid rounding errors in the
pixel pipeline while in hardware mode. These errors do not occur in the
software rasterizer. This workaround stretches the framebuffer so the
rounding error artifacts are not shown. However, this stretching creates
an improperly sized framebuffer which may cause other problems. I
recommend no longer using that patch except for the portions which do
not relate to stretching the framebuffer. I can post those code bits if
needed.

Instead, I identified the source of the rounding errors and posted a
patch below. This fixes the pixel pipeline in hardware mode and affects
both 3dfx and d3d games. The d3d games were especially susceptible to
the errors, likely from extensive use of the pixel pipeline in HUDs and
text on-screen.