Disable mipmapping for Voodoo software rasterizer
This change is specific to the Voodoo software rasterizer and effectively disables its mipmapping; however, the opengl path is unaffected. Where mipmapping is active, then textures are compressed to low resolution at a near distance whereas this should occur farther away. For GLQuake, the gl_nearest and gl_linear modes are available. UT does not offer the mipmap option in its preferences for glide mode, but this commit will workaround the above texture issue.