Patch: fix win32 mouse input latency in SDL 1.2hg
After the release of SDL 1.2.13, a commit for their bug# 611 was applied to enhance ioquake3 mouse input in Windows, particularly to enable dinput in an opengl application and for reliability of mouse input under the windib driver. For example (from their commit): "Googling for this indicates that often this is known to result in "spurious" and/or "missing" mouse movement events...". It also added a dependency on dinput7.

The attached patch reverts the above commit and it applies to a recent version of SDL 1.2hg (tested in mingw32 and Vista). This restores the mouse input latency to that of SDL 1.2.13 and fixes specific mouse cursor stuttering issues (limited testing of windib and directx under output=surface and =opengl in games such as SC2000, FS 5.1, and UT99/3dfx). It also fixes other minor mouse issues, such as the major loss of mouse movement events while the 3dfx emulation is active in the opengl or software rasterizer path. The previous workarounds