Revert "Let's just disable PulseAudio entirely in our SDL builds and be done"
This reverts commit b70a024a0806429b49ede0b86969be9aa74ce790.

Okay fine PulseAudio. I pulled in SDL 1.x for a reason, and here it is.
The audio thread handling needs to be fixed, just as the Windows builds
need to fix the SDL 1.x issue with relative mouse motion. This is why.
Time to start hacking. If it makes it incompatible with the main SDL 1.x
library at this point I don't care, it's time to fix this nonsense.