SVN r3962 and r3971
Set record size to 128 if it is 0 in the FCB when calling any FCB
read/write function
Fixes for acad 10:
- Add missing reference counting when the file is already open when
calling FCB_Open, so that acad, which uses FCBs and normal handles on
the same file, works better.
- Remove FCBs being added to the PSP filetable and rewrite most
functions to support this change. This way acad won't run out of
temporary (fcb) files when low on memory.