SVN r4197
Change multi-remain to repeat last parsed value for non-string types if
the next value is empty and of the same type as the last.

Use this to add an optional parameter to sensitivity which controls the
y axis. When optional parameter is missing, x and y axis have the same
value.
Change limits on sensitivity to allow for negative values so the Mouse
Y-axis can be inversed.