SVN r3980 (partial)
Faked successes for INT 13 format calls were already in DOSBox-X, but
the SVN implementation includes returning a failure when the drive is
inactive. Implemented that part.

The tracking of the last READ or WRITE from the SVN commit is omitted,
since the code this is used as a condition with is a little different in
DOSBox-X.