SVN r3961 (modified)
Add support for Print Screen key and interrupt. The key press allows
Descent and 2 in-game screenshot feature to be used.

Also, according to the original SVN commit text, "The keypress and
interrupt allow the Horror hotel (TSR text adventure) hotkey to work,
and the compatible interrupt handler location fixes a game bug in The
Forgotten Land."