Port phonebook feature from dosbox-staging https://github.com/dosbox-staging/dosbox-staging/commit/e905a6bd5d2f0d0d0fa07d904f9dab75a929ecbf
create a phoneboox-<version>.txt file with the layout:
<number> <host:port>

When you dial the number, it will actually connect to the specified host. This is needed
as not all programs allow to enter a hostname or IP address.

It also adds a new property to the config file in the [serial] section named phonebookfile=