Only use fopen64/etc on Linux and Windows
Use the standard variants of these functions everywhere else,
since other OSes have generally implemented large file support
on top of the standard functions.

There are other exceptions... Solaris and QNX are the only others
I could find with google that hint that they implement the *64
functions, but Solaris at least has deprecated 32-bit support,
which makes it a bit irrelevant. Unfortunately I can't test on
QNX because my QNX hardware is broken. I could be useful with
fixing dosbox-x on Solaris but I'm not sure how to get endian
definitions out of it (maybe SDL could be used for this, though).

This helps me with building on NetBSD.