Inform on unexpected control characters in batch files
This commit informs the user when an illegal character is
encounter in a batch file.  It does not change the actual
parsing or handling of the illegal character, which is to
skip the character.

Because some control characters are invisible when TYPE'd
on the command line or viewed in an editor, this messaging
will alert users that something is (very likely) amiss in
their batch file and could be the cause for unexpected
behavior or parsing.

patch by kcgen ported from dosbox-staging
https://github.com/dosbox-staging/dosbox-staging/commit/be07ddb03106dfe407796de6055cbedad0936bf8