Document country settings
No functional changes, just documentation to better understand what is happening.

Also, what is country 88 meant to be? It does not appear in any table I have been able to find. Generally the codes are based on the international dialling codes of countries, and there is no 88 dialling code as it would conflict with several 3-digit dialling codes starting with 88 (Bangladesh (880) and Taiwan (886)).