Change default Fluidsynth fluid.gain to 0.2
Upstream Fluidsynth development uses a gain of 0.2 by default.
For some reason the Fluidsynth in DOSBox-X (and DOSBox ECE) instead
uses a value of 0.6

This causes problems with games that use both MIDI and SoundBlaster
for speech and sound effects. The MIDI output will overpower the
SoundBlaster volume. Changing the gain to 0.2 fixes this.