Move Linux packaging files to contrib/linux
The general idea is to clean up the root directory of the tree
a bit, and to possibly prepare to support flatpak in the future
without adding more files to the root.

Both "make install" and make-rpm.sh worked after these changes.

My idea was also to move src/dosbox.png to contrib/icons/dosbox-x.png
And to move any Windows and macOS specific packaging files to
contrib/windows and contrib/macos respectively, but I will not do the
latter two myself as I cannot test it.

This is also similar to how dosbox-staging has it's tree layout.

If there is no objection, I will however move src/dosbox.png and
update any references to it to point to the new location.