Switch linux to SVG icon
I found that the icon used by DOSBox-X comes from
https://www.vogons.org/viewtopic.php?p=79575#p79575

Which apart from Windows and Mac icons also included a SVG icon
which is preferred on Linux, and looks nicer. So add the SVG icon
and update 'make install' and make-rpm.sh to use it.
The SVG icon is placed in contrib/icons

p.s. I found that the tar operation in make-rpm.sh blindly assumes
the directory containing the tree to be named "dosbox-x", instead
of just the current directory, it effectively will tar
../dosbox-x even if your in, say, dosbox-x.clone to keep different
trees apart... I did not change this for now, but it is something
to watch out for.