Fix AVX2 when not building with -mavx2 or similar
In generic builds, we have to tell the compiler to change the target for
a specific function and to make sure to call it only if we know the
extension is supported by the CPU.  This is achieved using this target
function.

I’ve tested this code on both gcc and clang, with -march=native and
without, with different optimisation levels, and it seems sound this
time, sorry for breaking the generic build.

From another project, MSVC doesn’t have this restriction.

Fixes #1880.
Fixes #1881.