Add screenshots for metainfo.xml
This adds some screenshots for metainfo, that is used for
flatpak and the software centre application on linux.

If you prefer the screenshots to be stored someplace else,
let me know. I could move them to a different directory such
as contrib/screenshots or to the gh-pages branch, or I could
conceivably put them into the flathub repository.