Add windowposition config option
Normally, at least on Windows, the window open with SDL1 is positioned
randomly on the screen, which is pretty annoying. It can be overriden
by setting SDL_VIDEO_WINDOW_POS, but it's cumbersome.

SDL2 is even worse, as SDL_VIDEO_WINDOW_POS does nothing and the window
always ends up in the same, weird place on the screen (top left corner
is around the middle of the screen).

Added "windowposition" option to allow users to configure the exact
placement of the window when DOSBox starts. If the config value is empty,
no change will be made.