Fix URL and resize images to 16/9
Old URL does not actually return the image, but a github page

In addition, "appstream-util validate" fails because the image is
not 16/9. And this is run as part of the flatpak build process,
which causes the build to fail.

So I resized the images to 16/9, by adding transparent horizonal
padding and centering the images.