Cleanup RPM spec file
I am working to submit dosbox-x to Fedora for possible inclusion,
but for that I need to ensure I had a suitable spec file. The
one included in the source tree had to be cleaned up.

This is still not 100% the same as the spec file I will use to
submit dosbox-x for a packaging review, but it is close.

One difference is, that this spec file generates a SDL1 binary,
while I will be submitting a SDL2 version upstream for 2 reasons
- SDL2 should be future
- Fedora is unlikely to accept the SDL1 version with customized SDL1

I can change the in-tree spec file to also generate the SDL2 version,
if there is no opposition.