Updated french language with chcp modification too.
Hello.

This is another big update version. Because of chcp modification, I had to modify my translation. Like in the previous version, help command are more pretty, PC-98 intro screen is OK, and I start working on mapper editor part.

There is a lot of modifications. This patch makes the french translation finished up to at least 85 to 90%. And this time, I did not touch codepage :)

@Wengier: please, merge it.

Thanks.