Fix 4 operator instruments not working on RetroWave OPL3
The RetroWave OPL3 handler used an opl3_port variable to record the last
register set that had received a register address write. When a register value
was written, the variable was used to determine which register set the value
should be written to. WriteAddr would just return the address that was written.

This approach has a problem when it comes to the timer registers which are
emulated by the Chip class. Two of the timer registers in the first register
set (04 and 05) have different functions on an OPL3 in the second register set
and must not be handled by the Chip class, but should be sent to the OPL3. For
the 05 register (which turns OPL3 features on and off) an exception was already
added, which would add an extra bit to the address returned by WriteAddr
indicating that the address was in the second register set. This prevented
Chip::Write from handling the register value write. The problem remained for
the 04 register, which is used to turn 4 operator channels on and off.

Instead of adding another exception, the bit indicating the second register set
is now added to all addresses written to the second I/O port. WriteReg now uses
this extra bit instead of opl3_port to determine which register set a value
should be written to. This fixes the 4 operator issue and also allows for
easier support of dual OPl2.