Enable OPL types other than OPL3 for RetroWave OPL3
This enables OPL2, dual OPL2 and OPL3 Gold modes for the RetroWave OPL3.

An OPL3 chip is in an OPL2 compatibility mode by default, so nothing is needed
to make this mode work. The only differences between SB OPL3 and AdLib Gold
OPL3 are the I/O addresses used and an extra controller chip, which are already
emulated in software. Note that the volume control is implemented using the
DOSBox mixer, so that will not work with the RetroWave OPL3.

Dual OPL2 is implemented using software OPL3 emulators, so most things needed
to make this work on an OPL3 are also already implemented. The only problem
when using a hardware OPL3 chip is that software using dual OPL2 will write to
register 01 on both chips (which is both register sets on an OPL3) to set bit 5
which enables waveform select on an OPL2. This does nothing in the first
register set on an OPL3 (waveform select is always enabled), but in the second
register set this bit is a test bit. OPL3 emulators don't implement this, but
setting this bit on a real OPL3 will disable sound output. To prevent this from
happening, writes to both 01 registers are disabled for the RetroWave OPL3.