fixes joncampbell123/dosbox-x#3175
Makes volume name matching use the standard 8.3 codepath.  This seems to be expected behavior as the volume for Stellar7 floppies is of the form
"Stellar7 1", however the installer expects to be able to search for both "Stellar7??" as well as "Stellar71" implying that trimming SHOULD be happening on both the name as well as the extension before wildcard matching even for volume labels.

This adds an additional variable vol_name to construct the intended output and cached volume label, but does not involve this variable in any wildcard matching.