ne2000: Trigger pending interrupts instead of panicing
When sending packets the NE2000 driver triggers an interrupt according
to some magic number that factors in the packet length, probably to
emulate the time it takes to process and actually send a packet.

Currently if an interrupt is already pending when sending a packet the
emulation will panic despite having code that just removes the pending
interrupt straight after. I'm guessing at some point this panic was just
a debug statement. Removing the panic seems to result in a working system.

kcgen from dosbox-staging already found and fixed this bug here:
https://github.com/dosbox-staging/dosbox-staging/commit/4a5c3fefb687f54937cf5433ef753fff8b5581a6

Their patch immediately triggers an interrupt if there's one pending
which makes more sense than dropping the interrupt so I copied that
idea for this patch.

Their patch also reduces the time it takes to send a packet by 10 in
hope of avoiding hitting a pending interrupt instead. I've chosen not to
copy this in case applications are relying on the current timing.

I don't have a real NE2000 to test this behaviour against, but I doubt
the card would crash in this situation.

Reported-by: gogoud from DOSBox-X discord
Tested-by: gogoud from DOSBox-X discord
Co-authored-by: kcgen <1557255+kcgen@users.noreply.github.com>