shell_cmds: Fix number overflow
For an output of any number of 1000GB or more 16 bytes is not enough to
represent it: "1,000,000,000,000" takes 18 bytes including the NULL
terminator. Increase the buffer size to something overkill like 64 so we
never have to think of this situation again.

Strangely this was found to crash MinGW builds on Windows but not on
other compilers or platforms. Shortening the numformat to 5 easily
reproduces the crash on my system.

Closes #3644 https://github.com/joncampbell123/dosbox-x/issues/3644

Reported-by: gkswngh05 (https://github.com/gkswngh05)
Fixes: bd7a7293194747ee9ae6492a1a5c5dec9a4b398a