debug: Don't consume trailing spaces in GetHexValue
With the change to add math expressions to GetHexValue it needs to parse
spaces in order to check if there's an operator. Unfortunately this
parses trailing spaces unconditionally even if there's no operator,
this breaks code that expects whitespace to be present after
GetHexValue's return.

For instance, 'MEMDUMP 0:0 1234' would skip the next character after
0:0 is parsed by GetHexValue. With a greedy consumption of spaces it
parses the 1 in 1234 and then dumps address 0x234.

Thanks to grapeli on GitHub for bisecting and finding the problematic
commit.

Closes #3629 (https://github.com/joncampbell123/dosbox-x/issues/3629)

Reported-by: AK5GH (https://github.com/AK5GH)
Fixes: b084c17b63f2d2e5d2b5d08c14237648e0f79ac9