fpu: Drop support for using 'long double' for 80-bit floats
The only systems I can find where 'long double' is 80-bit IEEE is
on x86 and Itanium. For x86's case it can use the FPU directly instead
of going through a structure. Itanium is dead and hard to support.

Having this code has already caused confusion on non-x86 platforms
and we already special-case ARM. Instead of special casing this to
be x86 and GCC only we might as well just dump the code entirely.

This also drops the glibc-specific fpu_control.h dependency.