Update the build-macos and build-macos-sdl2 scripts to build both arm64 and x86_64
Universal App Bundles require we build the executables twice. Update the build scripts to support building both arm64 and x86_64 architectures. By default the scripts should behave like they always did, despite the major overhaul, and only build for your host's architecture. To build *both* architectures, you *must* run the build on an Apple Silicon Mac *and* run the script with the word `universal` as the first parameter - any parameters thereafter are used as-is for the `configure` script of DosBox-X like the build script did before it supported the extra parameter.

Signed-off-by: Dani Llewellyn <dani@bowlhat.net>