MIDI: set minimum sysex delay when enabled
A real rev.0 MT-32 is the main system that needs the sysex delay, and
it needs about 40ms between sysex commands to work properly. 40ms is
documented in a few blogs online and is the delay used universally by
ScummVM.

Reviewing the original delaysysex patch, the case that's leading to
incorrect timing here is a new case in the if statement that wasn't in
the original patch. It's possible this bug was introduced later, or that
the patch as written wasn't entirely comprehensive enough.

Tested with MEdit, which was written for LAPC-I/CL-32L; with this patch,
loading problematic songs (such as One Step Beyond's INGAME1.RLD) now
works, whereas they reliably triggered a buffer overflow before.